Add updater safety-path tests and testable helper module
Extract pure installer parsing/asset-selection helpers and cover

version parsing, asset filtering, checksum matching, and hash parsing.
