+++ package.json
@@ -498,6 +498,7 @@
     "vscode:prepublish": "npm run compile",
     "compile": "tsc -b",
     "watch": "tsc -b -w",
-    "lint": "eslint ./src --ext .ts,.tsx"
+    "lint": "eslint ./src --ext .ts,.tsx",
+    "test": "npm run compile && node --test test/**/*.test.js"
   }
 }
